requirements.txt
PyYAML
Pillow
mapproxy
